requirements.txt
python-telegram-bot>=21.0
requests>=2.31.0
APScheduler>=3.10.4
python-dotenv>=1.0.0
pytz>=2024.1